+++ search-films/CHANGELOG.md
@@ -15,8 +15,9 @@
 - add counter of favour films (e59887a1)
 - **user.slice:** add user.slice (need to test) (ea751c63)
 - **MovieCard:**
-  - impl of add/remove from favour (5f9266a3)
-  - add util component MovieCard, minimal refactoring (a652f0d6)
+  - add/delete WishList (2f8cd391)
+  * impl of add/remove from favour (5f9266a3)
+  * add util component MovieCard, minimal refactoring (a652f0d6)
 
 #### 0.3.1 (2024-11-13)